# api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Satoru Anime API",
    "version": "1.0.0",
    "description": "API for scraping and retrieving anime information from Satoru"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/search": {
      "get": {
        "summary": "Search for anime",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "Search term",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful search results",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "results": [
                      {
                        "id": 123,
                        "title": "example-anime",
                        "image": "https://example.com/image.jpg",
                        "type": "TV"
                      }
                    ],
                    "currentPage": 1,
                    "hasNextPage": true,
                    "totalResults": 50
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/info/{animeId}": {
      "get": {
        "summary": "Get anime information",
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "ID of the anime",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime information",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": "123",
                    "title": "Example Anime",
                    "image": "https://example.com/image.jpg",
                    "description": "Anime description",
                    "genres": ["Action", "Adventure"],
                    "status": "Ongoing",
                    "type": "TV",
                    "rating": "PG-13",
                    "releaseDate": "2023"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/episodes/{animeId}": {
      "get": {
        "summary": "Get episodes list",
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "ID of the anime",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Episodes list",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "episodes": [
                      {
                        "id": "ep1",
                        "number": 1,
                        "title": "Episode 1",
                        "japaneseTitle": "エピソード 1"
                      }
                    ],
                    "totalEpisodes": 12
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/servers/{episodeId}": {
      "get": {
        "summary": "Get available servers for an episode",
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "description": "ID of the episode",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server list",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "id": "server1",
                      "name": "Server 1",
                      "type": "direct"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/sources/{animeId}/{episodeId}": {
      "get": {
        "summary": "Get video sources for an episode",
        "parameters": [
          {
            "name": "animeId",
            "in": "path",
            "required": true,
            "description": "ID of the anime",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "description": "ID of the episode",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Video sources",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": [
                    {
                      "url": "https://example.com/video.m3u8",
                      "quality": "auto",
                      "isM3U8": true
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/id/{title}": {
      "get": {
        "summary": "Get Satoru Anime ID by title",
        "parameters": [
          {
            "name": "title",
            "in": "path",
            "required": true,
            "description": "URL-friendly title of the anime",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anime ID",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "id": 123
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Error message"
          }
        }
      }
    }
  }
} 